Parallelization
The possibility to run multiple datasets in parallel has been implemented. The 'input_data_list' parameter now indicates whether a single dataset or
or multiple datasets will be run at the same time. A new 'input_primaryBeam' parameter has been included to input the primary beams files if desired.
The main function has been adjusted as a try-finally block so that the .log file is reorganized in any case so that it is readable.
At the moment, core control is limited for SoFiA to only 1 per set regardless of what is indicated by terminal or in sofia.par, discussion and testing is needed.
Minor bugs have been fixed.