Parser input_data_set
The ability to enter multiple datasets at the same time directly in the config.yaml, either as a dict, string or list.
e same time directly in the config.yaml, either as a dict, string or list. The ability to add datasets via a .txt file has been removed for the time being.
This feature will be added back in future commits.