Parser input_data through a text file
Now data files can be passed via a text file which has the same logic as input_data_set. For this purpose a new parameter 'input_file' has been used.
This one is a bit more restrictive and a priori is not designed to work with lists but with dict-like lines like: 1 : files...