SIP paralelization bug fixed. Updates and bugs fixed
SIP can now run in parallel even if SoFiA and TAP are disabled. Additionally, if TAP is enabled and SoFiA is not,
SIP will attempt to search for possible existing catalogs based on the names of downloaded files and default
directories. The README has been partially updated and some bugs fixed. Minor updates have been made to certain
parts of the code