SoFia log file added to adpalmap log file.
In the sopar.py module, variables have been added that are returned to the process_data function and from there they are saved in worker_result.
These contain the necessary information to associate each line of the pipeline logger:

2025-05-16 12:00:27,435 | INFO | [PID:12263] sopar: - SoFia start. Mode: absorption. Input data: {filename} (in which the Input data information had to be added)

with the SoFia .log file. In this way, after this line, everything included in each .log file will be included in the final pipeline logger.