Adding SIP logger in progress. Par. -log_filename.
Changes have been added to handling the SIP logger file to reorganize and append it to the final ADPALMAP .log file.
The SIP -log_filename parameter has been added to the SIP command, but it cannot be included or handled if added as a
SIP parameter; it will return an error indicating that it does not exist.