Fix mask conversion bug and add a parameter to delete float masks
Fixes a bug where mask were transformed from float to int and overwrite the file. This cause ADPALMAP
to detect in subsequent runs that file as a incomplete download (1/2 in size exactly).
The download is attempted to complete resulting in a corrupted file and causing the error

Also adds a new parameter in download_par to allow deleting the original
float-format mask file after processing.